requirements.txt
pandas
polars
numpy
tqdm
jinja2
pydantic
lightgbm
click
streamlit
openai
python-dotenv
